Moved the location of these files to make it more suitable for puckage structure.
